+++ noodles-sam/CHANGELOG.md
@@ -500,7 +500,7 @@
 
   * sam/lazy: Add lazy record (`lazy::Record`).
 
-    Lazy records are alignment records that are lazily-evaluated. They fields
+    Lazy records are alignment records that are lazily-evaluated. Their fields
     are not necessarily valid, but the buffer is guaranteed to be record-like.
 
   * sam/reader: Add `Reader::read_lazy_record` to read lazy records.